+++ mobile/composeApp/src/commonMain/kotlin/com/yagubogu/ui/ranking/model/RankingProfileItem.kt
@@ -2,14 +2,14 @@ package com.yagubogu.ui.ranking.model
 
 sealed interface RankingProfileItem {
     val memberId: Long
-    val rank: Int
+    val rank: Long
     val nickname: String
     val profileImageUrl: String
     val teamName: String
 
     data class CheckInRanking(
         override val memberId: Long = 0L,
-        override val rank: Int = 0,
+        override val rank: Long = 0,
         override val nickname: String = "",
         override val profileImageUrl: String = "",
         override val teamName: String = "",
@@ -18,7 +18,7 @@ sealed interface RankingProfileItem {
 
     data class VictoryFairyRanking(
         override val memberId: Long = 0L,
-        override val rank: Int = 0,
+        override val rank: Long = 0,
         override val nickname: String = "",
         override val profileImageUrl: String = "",
         override val teamName: String = "",
